requirements.txt
smbus-cffi
python-ev3
itsdangerous
Flask
Flask-RESTful
Flask-Login
Flask-Principal
